Rename stack to space, make invokePlugin output consistent
