+++ __tests__/plugins/divideWidth.test.js
@@ -20,24 +20,16 @@ test('generating divide width utilities', () => {
 
   expect(utilities).toEqual([
     [
-      [
-        {
-          '.divide-y-2 > * + *': { 'border-top-width': '2px' },
-          '.divide-x-2 > * + *': { 'border-left-width': '2px' },
-        },
-        {
-          '.divide-y-4 > * + *': { 'border-top-width': '4px' },
-          '.divide-x-4 > * + *': { 'border-left-width': '4px' },
-        },
-        {
-          '.divide-y-8 > * + *': { 'border-top-width': '8px' },
-          '.divide-x-8 > * + *': { 'border-left-width': '8px' },
-        },
-        {
-          '.divide-y > * + *': { 'border-top-width': '1px' },
-          '.divide-x > * + *': { 'border-left-width': '1px' },
-        },
-      ],
+      {
+        '.divide-y-2 > * + *': { 'border-top-width': '2px' },
+        '.divide-x-2 > * + *': { 'border-left-width': '2px' },
+        '.divide-y-4 > * + *': { 'border-top-width': '4px' },
+        '.divide-x-4 > * + *': { 'border-left-width': '4px' },
+        '.divide-y-8 > * + *': { 'border-top-width': '8px' },
+        '.divide-x-8 > * + *': { 'border-left-width': '8px' },
+        '.divide-y > * + *': { 'border-top-width': '1px' },
+        '.divide-x > * + *': { 'border-left-width': '1px' },
+      },
       ['responsive'],
     ],
   ])
